# app/Http/Resources/GradeResource.php
<?php

namespace App\Http\Resources;

use Illuminate\Http\Request;
use Illuminate\Http\Resources\Json\JsonResource;

class GradeResource extends JsonResource
{
    /**
     * Transform the resource into an array.
     *
     * @return array<string, mixed>
     */
    public function toArray(Request $request): array
    {
        return[
            'id'=> $this->id,
            'student_id'=> $this->student_id,
            'subject' => SubjectResource::make($this->whenLoaded('subject')),
            'score'=> $this->score,
            "exam_date"=>$this->exam_date,
        ];
    }
}