# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Система оценок (Gradebook API)",
        "description": "API для управления студентами, предметами и их оценками. Использованы Laravel Resources и hasManyThrough.",
        "contact": {
            "email": "artemtaiti@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Laravel API Server"
        }
    ],
    "paths": {
        "/students/{student}/grades": {
            "get": {
                "tags": [
                    "Оценки"
                ],
                "summary": "Получить все оценки студента",
                "description": "Возвращает отсортированный список оценок по ID студента, включая предмет (subject).",
                "operationId": "0a7de7d3170cca4e49baedbd5ba5747c",
                "parameters": [
                    {
                        "name": "student",
                        "in": "path",
                        "description": "ID студента",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/GradeResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Студент не найден"
                    }
                }
            },
            "post": {
                "tags": [
                    "Оценки"
                ],
                "summary": "Создать новую оценку для студента",
                "operationId": "499dd58a60faa1a91b04cf77fc06b3db",
                "parameters": [
                    {
                        "name": "student",
                        "in": "path",
                        "description": "ID студента",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "subject_id",
                                    "score",
                                    "exam_date"
                                ],
                                "properties": {
                                    "subject_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "score": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "exam_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-11-20"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Оценка успешно создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/GradeResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Студент не найден"
                    }
                }
            }
        },
        "/students/{student}/grades/{grade}": {
            "put": {
                "tags": [
                    "Оценки"
                ],
                "summary": "Обновить оценку",
                "operationId": "30e2173d741ab8ba09a3afa089eaaa75",
                "parameters": [
                    {
                        "name": "student",
                        "in": "path",
                        "description": "ID студента (для проверки принадлежности)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "grade",
                        "in": "path",
                        "description": "ID оценки",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "subject_id",
                                    "score",
                                    "exam_date"
                                ],
                                "properties": {
                                    "subject_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "score": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "exam_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-11-20"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Оценка успешно обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/GradeResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Оценка не принадлежит указанному студенту"
                    },
                    "404": {
                        "description": "Студент или Оценка не найдены"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Оценки"
                ],
                "summary": "Удалить оценку",
                "operationId": "402813e57ab6930a1f8c6aa3aac55d11",
                "parameters": [
                    {
                        "name": "student",
                        "in": "path",
                        "description": "ID студента (для проверки принадлежности)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "grade",
                        "in": "path",
                        "description": "ID оценки",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Оценка успешно удалена"
                    },
                    "403": {
                        "description": "Оценка не принадлежит указанному студенту"
                    },
                    "404": {
                        "description": "Студент или Оценка не найдены"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "GradeResource": {
                "title": "Grade Resource (Output)",
                "description": "Оценка, возвращаемая API, включая связанный предмет",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "student_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "score": {
                        "type": "integer",
                        "example": 5
                    },
                    "exam_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-11-20"
                    },
                    "subject": {
                        "description": "Связанный предмет",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 3
                            },
                            "title": {
                                "type": "string",
                                "example": "Высшая математика"
                            },
                            "lecturer_name": {
                                "type": "string",
                                "example": "Петров"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}